pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "seo-farm-orchestrator"
version = "1.0.0"
description = "AI-powered SEO content orchestrace pomocí Temporal.io + OpenAI Assistant API"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "SEO Farm Orchestrator", email = "admin@seo-farm.com"}
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.11"
dependencies = [
    "temporalio==1.4.0",
    "httpx==0.27.0", 
    "pydantic==2.7.1",
    "python-dotenv==1.0.0",
    "openai==1.58.1"
]

[project.optional-dependencies]
db = [
    "sqlalchemy>=2.0.23",
    "psycopg2-binary>=2.9.9"
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0"
]

[project.scripts]
seo-farm = "seo_farm_orchestrator.cli:main"

[project.urls]
Homepage = "https://github.com/your-org/seo-farm-orchestrator"
Repository = "https://github.com/your-org/seo-farm-orchestrator"
Documentation = "https://github.com/your-org/seo-farm-orchestrator#readme"

[tool.setuptools.packages.find]
where = ["."]
include = ["seo_farm_orchestrator*"]

[tool.setuptools.package-data]
seo_farm_orchestrator = ["*.py"] 